third_party/WebKit/LayoutTests/fast/text/soft-hyphen-5-expected.txt

Soft Hyphen Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS true is true
PASS successfullyParsed is true

TEST COMPLETE
anti­dis­est­ab­lish­ment­arian­ism.
